Revert "LPD-37169 portal-search: Add & update tests"
This reverts commit a4176e32247a03ab16c48691a679c4088b8284b1.
